refactor(test): 重构test目录
