+++ tests/e2e/__init__.py
@@ -1,0 +1,1 @@
+"""Isolated end-to-end support modules."""
