+++ tests/e2e/admin/__init__.py
@@ -1,0 +1,1 @@
+"""Administrator end-to-end support modules."""
