+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Cross-module integration contract tests."""
