+++ tests/integration/admin/test_admin_deployment.py
