+++ tests/unit/admin/__init__.py
@@ -1,0 +1,1 @@
+"""Administrator unit tests."""
