+++ tests/unit/admin/test_admin_cli.py
