+++ tests/unit/admin/test_admin_settings_api.py
