+++ tests/unit/auth/test_stale_session_guard.py
