+++ tests/unit/database/__init__.py
@@ -1,0 +1,1 @@
+"""Database unit tests."""
