+++ tests/unit/database/test_monitor_settings.py
